pyproject.toml
[project]
name = "coxireduxer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "llama-cpp-python>=0.3.13",
    "ollama>=0.5.1",
    "tqdm>=4.67.1",
]
